Eliminación de archivos simples y mover lógica
