+++ src/database/connection.ts
@@ -5,13 +5,3 @@ const { Pool } = pkg;
 export const pool = new Pool({
   allowExitOnIdle: true,
 });
-
-// try {
-//   // await pool.query("SELECT NOW()");
-//   pool.query("SELECT NOW()");
-//   console.log("Database connected");
-  
-// } catch (error) {
-//   console.log(error);
-  
-// }
