--- src/helpers/convertToLowercase.ts
@@ -1,3 +1,0 @@
-export const lowercase = (text: string) => {
-  return text.toLowerCase();
-};
