+++ src/helpers/sendEmail.ts
@@ -1,15 +1,13 @@
-import { Response } from "express";
 import * as nodemailer from "nodemailer";
-import { resolve } from "path";
 
 const transporter = nodemailer.createTransport({
-  service: "Gmail", // o cualquier otro servicio de correo que utilices
+  service: "Gmail",
   host: "smtp.gmail.com",
   port: 465,
   secure: true,
   auth: {
-    user: process.env.MAIL_USER, // tu correo
-    pass: process.env.MAIL_PASSWORD, // tu contraseña
+    user: process.env.MAIL_USER,
+    pass: process.env.MAIL_PASSWORD,
   },
 });
 
@@ -19,22 +17,19 @@ export const sendEmail = (
   name: string,
   token?: string
 ) => {
-  // <div style="text-align:center;background:#192c7b;">
-  //   <img src="cid:uniqueImageCID" alt="Logo CCIOD Technologies" />
-  // </div>;
-  let content = `
+  let html = `
     <html lang="es">
       <head>
         <title>Reset Password</title>
       </head>
       <body>
         <div style="color: #121519; text-align:justify;">`;
-  content += `<p>Hola, <span style="font-weight: bold;">${name}</span>, `;
-  content += token
+  html += `<p>Hola, <span style="font-weight: bold;">${name}</span>, `;
+  html += token
     ? `hemos recibido una petición para reestablecer su contraseña.</p>`
     : `su contraseña se ha reestablecido.</p>`;
-  content += `<hr style="opacity: .20; margin: 1rem 0;" />`;
-  content += token
+  html += `<hr style="opacity: .20; margin: 1rem 0;" />`;
+  html += token
     ? `<p>Haz clic en el botón de abajo y crea una nueva contraseña. Si no solicitaste restablecer tu contraseña, ignora este correo electrónico. En caso de que hayas recibido más de un correo, debes utilizar el último que te llegó.</p>
       <div style="text-align:center; margin-top: 2rem;">
         <a href="${process.env.FRONTEND_URL}reset-password/${token}" style="background: #192c7b; color: #eee; margin: 1rem 0;padding: .8rem; border-radius: .5rem; cursor: pointer; text-decoration:none;">Reestablecer contraseña</a>
@@ -43,7 +38,7 @@ export const sendEmail = (
           <div style="text-align:center; margin-top: 2rem;">
         <a href="${process.env.FRONTEND_URL}" style="background: #192c7b; color: #eee; margin: 1rem 0;padding: .8rem; border-radius: .5rem; cursor: pointer; text-decoration:none;">Iniciar sesión</a>
       </div>`;
-  content += `<hr style="opacity: .20; margin: 1rem 0;" />
+  html += `<hr style="opacity: .20; margin: 1rem 0;" />
           <div style="text-align:center;">
           <span>© 2024 CCIOD,  Todos los derechos reservados.</span>
           </div>
@@ -54,14 +49,7 @@ export const sendEmail = (
     from: process.env.MAIL_USER,
     to,
     subject,
-    html: content,
-    // attachments: [
-    //   {
-    //     filename: "cciod-technologies.png",
-    //     path: __dirname + "../../assets/cciod-technologies.png",
-    //     cid: "uniqueImageCID", // Referenced in the HTML template
-    //   },
-    // ],
+    html,
   };
   return new Promise((resolve, reject) => {
     transporter.sendMail(mailOptions, (err, response) => {
@@ -70,17 +58,6 @@ export const sendEmail = (
       } else {
         resolve(response);
       }
-      // if (err) {
-      //   return res.status(400).json({
-      //     success: false,
-      //     message: "Ocurrió un error al intentar enviar el correo.",
-      //   });
-      // } else {
-      //   return res.status(201).json({
-      //     success: true,
-      //     message: "Correo enviado correctamente.",
-      //   });
-      // }
     });
   });
 };
